Fixed test_uttp after latest additions to the UTTP protocol.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@52737 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/test/test_uttp.cpp
@@ -291,7 +291,7 @@ bool CUTTPTestApp::TestWriter(size_t buffer_size)
 
 int CUTTPTestApp::Run()
 {
-    size_t buffer_size = (sizeof(size_t) >> 1) * 5 + 1;
+    size_t buffer_size = (sizeof(Int8) >> 1) * 5 + 1;
 
     do
         if (!TestWriter(buffer_size))
